1.reformat all the cmake files to be more readable
2.updating cmake to enable lto

3.added stricter warnings and stack checking

4.modified file to conform to the new strict flags

5.updated cmakelists to use abs paths

+++ include/cu/debugtimer.h
@@ -181,7 +181,7 @@ struct PrintTimeBlock : TimeBlock{
         GetTime(&end);
         
         auto len = GetTimeDifferenceMS(start_stamp,end);
-        printf("%s %s %d : %f\n",file,function,line,len);
+        printf("%s %s %d : %f\n",file,function,line,(f64)len);
     }
 };
